Added a quickpickmenu to choose folders  to display
The provided choice are, when relevant  :
* current document folder
*  home folder
* workspace root folders

There is a regression. The quickinput menu allowed to
choose any folder. I don't know if the quickpick menu
can double as a quickinput, meaning accepting the typed
string when no match. If not, one needs an additional entry
to the quickpick menu to launch a quickinput one.
Probably will go with that second option if I am not figure
if the second is possible.